Initial Technical Issues

+++ Window.Application/Services/Interfaces/IInquiryService.cs
@@ -22,7 +22,7 @@ namespace Window.Application.Services.Interfaces
 
         Task<List<InquiryViewModel>?> ListOfInquiry(string userMacAddress);
 
-        Task<int?> InitialTotalSamplePrice(ulong brandId, ulong sampleId, int height, int width, int productCount , int? katibeSize, ulong userId , ulong glassId);
+        Task<double?> InitialTotalSamplePrice(ulong brandId, ulong sampleId, int height, int width, int productCount , int? katibeSize, ulong userId , ulong glassId);
 
         Task<int?> InitializeSamplesPrice(List<Sample?> samples, User user, int height, int width);
